Refactor: pkg/util into sub-pkgs
