+++ pkg/kubelet/BUILD
@@ -92,9 +92,9 @@ go_library(
         "//pkg/kubelet/volumemanager:go_default_library",
         "//pkg/security/apparmor:go_default_library",
         "//pkg/securitycontext:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/util/dbus:go_default_library",
         "//pkg/util/exec:go_default_library",
+        "//pkg/util/file:go_default_library",
         "//pkg/util/io:go_default_library",
         "//pkg/util/iptables:go_default_library",
         "//pkg/util/mount:go_default_library",
