+++ pkg/kubemark/BUILD
@@ -31,13 +31,13 @@ go_library(
         "//pkg/proxy:go_default_library",
         "//pkg/proxy/config:go_default_library",
         "//pkg/proxy/iptables:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/util/exec:go_default_library",
         "//pkg/util/io:go_default_library",
         "//pkg/util/iptables:go_default_library",
         "//pkg/util/mount:go_default_library",
         "//pkg/util/node:go_default_library",
         "//pkg/util/oom:go_default_library",
+        "//pkg/util/pointer:go_default_library",
         "//pkg/util/sysctl:go_default_library",
         "//pkg/volume/empty_dir:go_default_library",
         "//pkg/volume/secret:go_default_library",
