+++ pkg/master/tunneler/BUILD
@@ -25,7 +25,7 @@ go_library(
     tags = ["automanaged"],
     deps = [
         "//pkg/ssh:go_default_library",
-        "//pkg/util:go_default_library",
+        "//pkg/util/file:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/clock:go_default_library",
