+++ pkg/printers/internalversion/BUILD
@@ -32,7 +32,7 @@ go_test(
         "//pkg/client/clientset_generated/internalclientset/fake:go_default_library",
         "//pkg/kubectl/testing:go_default_library",
         "//pkg/printers:go_default_library",
-        "//pkg/util:go_default_library",
+        "//pkg/util/pointer:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/api/extensions/v1beta1:go_default_library",
