+++ pkg/util/file/BUILD
@@ -5,18 +5,23 @@ licenses(["notice"])
 load(
     "@io_bazel_rules_go//go:def.bzl",
     "go_library",
-    "go_test",
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["json.go"],
+    srcs = ["file.go"],
     tags = ["automanaged"],
 )
 
-go_test(
-    name = "go_default_test",
-    srcs = ["json_test.go"],
-    library = ":go_default_library",
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
     tags = ["automanaged"],
 )
