+++ pkg/util/pointer/pointer_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2014 The Kubernetes Authors.
+Copyright 2017 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,21 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package pointer
 
 import (
 	"testing"
-
-	"k8s.io/apimachinery/pkg/util/diff"
 )
-
-func TestStringDiff(t *testing.T) {
-	diff := diff.StringDiff("aaabb", "aaacc")
-	expect := "aaa\n\nA: bb\n\nB: cc\n\n"
-	if diff != expect {
-		t.Errorf("diff returned %v", diff)
-	}
-}
 
 func TestAllPtrFieldsNil(t *testing.T) {
 	testCases := []struct {
