+++ pkg/util/template/template.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2015 The Kubernetes Authors.
+Copyright 2017 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package template
 
 import (
 	"bytes"
@@ -41,6 +41,7 @@ func ExecuteTemplate(w io.Writer, templateText string, data interface{}) error {
 	return t.Execute(w, data)
 }
 
+// ExecuteTemplateToString executes templateText with data and output written to string.
 func ExecuteTemplateToString(templateText string, data interface{}) (string, error) {
 	b := bytes.Buffer{}
 	err := ExecuteTemplate(&b, templateText, data)
