+++ staging/src/k8s.io/apimachinery/pkg/util/diff/diff_test.go
@@ -86,3 +86,11 @@ object.A:
 		}
 	}
 }
+
+func TestStringDiff(t *testing.T) {
+	diff := StringDiff("aaabb", "aaacc")
+	expect := "aaa\n\nA: bb\n\nB: cc\n\n"
+	if diff != expect {
+		t.Errorf("diff returned %v", diff)
+	}
+}
