+++ test/e2e/framework/BUILD
@@ -70,8 +70,8 @@ go_library(
         "//pkg/kubelet/util/format:go_default_library",
         "//pkg/master/ports:go_default_library",
         "//pkg/ssh:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/util/exec:go_default_library",
+        "//pkg/util/file:go_default_library",
         "//pkg/util/labels:go_default_library",
         "//pkg/util/system:go_default_library",
         "//pkg/util/version:go_default_library",
