+++ test/e2e/workload/BUILD
@@ -38,7 +38,7 @@ go_library(
         "//pkg/controller/replication:go_default_library",
         "//pkg/kubectl:go_default_library",
         "//pkg/master/ports:go_default_library",
-        "//pkg/util:go_default_library",
+        "//pkg/util/pointer:go_default_library",
         "//plugin/pkg/scheduler/schedulercache:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/utils:go_default_library",
